feat: add new attribute to make a fallback cache value

+++ src/core/common/decorators/cache-remove/cache-remove.decorator.ts
@@ -2,6 +2,7 @@ import { Action, Input } from "@core/common/types";
 import { createCacheDecorator } from "../create-cache-decorator";
 import { ICacheRepository } from "@database/interfaces/cache.interface";
 import { createHashedKey } from "@core/utils/create-hash.util";
+import { CachePrefixEnum } from "@core/enums";
 
 /**
  * Decorator that removes cache entries by based on predefined conditions.
@@ -42,9 +43,9 @@ export function CacheRemove<I = { [key: string]: any }>(input: Omit<Input<I>, 't
    */
   const action: Action<I, any, ICacheRepository> = async ({ instance, key, method, args, repository, }) => {
     const output = await method.apply(instance, args);
-    await repository.remove({
-      key: input.hashable_key ? createHashedKey(key) : key as string,
-    });
+    const use_key = input.hashable_key ? createHashedKey(key) : key as string;
+    await repository.remove({ key: use_key })
+    if (input.fallback) await repository.remove({ key: `${CachePrefixEnum.FALLBACK}/${use_key}` });
     return output;
   };
